+++ view/frmCRUDSQL.cs
@@ -125,7 +125,7 @@ namespace WindowsForms
                     usuarioService
                         .Alterar(new Usuario()
                         {
-                            Id= Int64.Parse(txtId.Text),
+                            Id= Int32.Parse(txtId.Text),
                             Nome = txtNome.Text,
                             Email = txtEmail.Text,
                             Ativo = ckbAtivo.Checked,
